x11-drivers/xf86-input-evdev: Use monotonic timestamps for input events
This patch adds the support for xf86-input-evdev to turn on monotonic
timestamps if kernel supports it.

And it corrects a previous similar patch by setting the clock source
at the right place in EvdevOn() instead of EvdevCache(). EvdevCache()
is only called once when X registers the input device fd with kernel,
and if later X closes and reopens the input device fd (e.g. when the
system goes through suspend/resume cycle), the clock source setting
will be lost. EvdevOn() is the right place to set clock source since
it is called whenever X wants to open the input device and use it.

BUG=chrome-os-partner:12187
TEST=On Link, grep "time stamps" /var/log/Xorg.0.log
     Make sure monotonic timestamp is used for evdev
     Run "xinput test-xi2" and move fingers on touch device
     Note that valuator 4 (Touch Timestamp) is using
     monotonic timestamp.
     Suspend and resume the device and make sure the timestamp
     does not change back to real world timestamp.

Change-Id: Ia8618475960f3b63a0e3cb83ef61fc43cf0f2092
Reviewed-on: https://gerrit.chromium.org/gerrit/33247
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>
Commit-Ready: Yufeng Shen <miletus@chromium.org>