cros-binary: move openssh/wget depends to hard-host-depends
We need openssh/wget on the host development system because we execute
them.  We don't want them in the target SYSROOT (at least, not because
of this eclass).  So drop them from the cros-binary eclass DEPEND as
that is for target packages and move to the hard-host-depends RDEPEND.

BUG=chromium-os:19140
TEST=`cbuildbot {{x86,amd64}-generic,daisy}-paladin {amd64,arm,x86}-generic-full stout-release` passed
TEST=`./build_packages virtual/chromeos-firmware` no longer installs openssh/wget

Change-Id: I27b710eea3c89e75516ebe6e163c402b642f4a7b
Reviewed-on: https://gerrit.chromium.org/gerrit/33272
Reviewed-by: Brian Harring <ferringb@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>