Prevent sandbox build contamination
Sometimes it is convenient to build vboot_reference locally for quick
tests, just by running the 'make' command. This causes the ./build
directory created with all the generated output. If one tries emerging
vboot_reference-firmware after that, the build directory is copied
into the sandbox, and then 'make' in sandbox gets confused, leaving
the previously built objects for use.

This in turn causes a lot of subtle failures (for instance 32/64 bit
incompatibility when building u-boot).

A way around the problem is to run 'make clean' before actually
building the package.

BRANCH=none
BUG=none
TEST=manual
   . while in chroot run 'make' in src/platform/vboot_reference
   . cros_workon-link vboot_reference-firmware
   . emerge-link vboot_reference-firmware chromeos-u-boot

   this sequence of commands was failing before the fix and is
   succeeding now

Change-Id: I19119e4b761493dea358674917f27740850c7a7b
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34061
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>