x11-drivers/xf86-input-evdev: convert all patches to git diff format
It is already difficult to review diff's of diffs.  It is even harder when
the diff format changes due to the use of different diff'ing programs.

Converting them all to git --diff makes it easier to rebase these changes
going forward as we uprev the server.

** This patch should introduce no actual changes to the resulting binary

BUG=chromium-os:34695
TEST=xf86-input-evdev builds clean

Change-Id: I59d7a4e0324ac121c7122eaefca84bbd0f6a1b35
Reviewed-on: https://gerrit.chromium.org/gerrit/33886
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>