google-breakpad: Build target minidump-2-core and host minidump-2-core.32
We were only building minidump-2-core for use in the chroot on the host.
The problem is that minidump-2-core can only work on a minidump of the same
ELF format as how it was built, which does not bode well for ARM minidumps
on x86 hosts.  So put a target-built minidump-2-core in the image so
developers can run it directly on an ARM machine.

Also, build a 32-bit version of minidump-2-core for the chroot, and install
it as /usr/bin/minidump-2-core.32.  This is the same thing as what we do for
dump_syms.32.

BUG=chromium-os:34938
TEST=emerge-daisy: 'file /build/daisy/usr/bin/minidump-2-core'
     sudo emerge: 'file /usr/bin/minidump-2-core.32'

Change-Id: Iebe83f8135d3444d79f9ec8fc9372caa0c7213e9
Reviewed-on: https://gerrit.chromium.org/gerrit/34565
Tested-by: Michael Krebs <mkrebs@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Michael Krebs <mkrebs@chromium.org>