x11-drivers/xf86-input-evdev: Use Abs Dbl timestamps and other cleanup
This is the first of a set patches to clean up the scroll wheel
acceleration implementation.

This patch-to-a-patch converts the valuators from the older
"contiguous, integer" valuator style to the new "disjoint, double"
valuators used by CMT.  In fact, the name and format of the
REL_WHEEL/REL_WHEEL and START/END Timestamp valuators are all changed
to match the same valuators used by CMT for smooth scrolling.

To make use of the double valuators, the pixel calculation is modified
slightly to return a double which includes the fractional scroll amount.

Also, the definition and initialization of the valuators and their label
atoms is cleaned up a bit.  Instead of trying to reuse the real axis label
arrays, which should be reserved for kernel-defined axes, we keep track
of the begin and end timestamp valuator indeces during initialization,
so we can use them later to apply the timestamps to the smooth scrolling
motion events.

Typically, I would split this patch into several smaller changes, but it
is really difficult to create patch-of-patches, so I am just smashing this
all together.

BUG=chromium-os:34699
TEST=Verify scroll up/down/left/right still scroll properly for USB mice
TEST=Using xinput test-xi2, verify that the scroll wheel and timestamp
     valuators are being set correctly as doubles

Change-Id: Ic876feb0a724c693781bc0235fd7799abb43c898
Reviewed-on: https://gerrit.chromium.org/gerrit/34257
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>