x11-drivers/xf86-input-evdev: Refactor touch timestamp
Be a little more careful when initializing the "Touch Timestamp" valuator
and its label's atom.

This axis only is only initialized when the device emits touch events.
And, since it is appended to touch events' touchMask when they are being
posted to the X server, we only need to add it to the touchMask valuator
mask, not to the mt_mask & last_mt_vals masks.

Also, initialize the valuator by giving it a big range.

BUG=chromium-os:34699
TEST=Verify touch events have correct touch timestamp

Change-Id: I142587e5feed98e956213288664494dc02b92e8e
Reviewed-on: https://gerrit.chromium.org/gerrit/34269
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>