x11-drivers/xf86-input-evdev: remove evdev-2.6.99-fix-ms-mice.patch
This was originally a workaround for a mouse that exposed an absolute
valuator and caused our software stack to use absolute valuators for
all valuators. It's no longer needed, possibly b/c we ignore absolute
valuators.

BUG=chromium-os:35178
TEST=Tested that MS Wireless Mouse 3500 scrolled, moved, clicked
correctly. This was the mouse that caused us to put this patch in
originally.

Change-Id: Ib37523fe62347cf0e7c170caeed8440cd3ac0619
Reviewed-on: https://gerrit.chromium.org/gerrit/35110
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>