Propagate the svn environment into clang 32-bit bootstrap
The chroot bootstrap process involves subversion checkout
within a sandboxed environment.  The sandboxed environment
did not include inheritance of the user's proxy settings for
subversion into the ${HOME} directory where they will be
expected if svn is invoked directly.  The 32-bit build of
the asan library of clang does this, with the resulting
symptom that the 32-bit build of clang's asan library failed
to checkout when run from behind a firewall.

This change propagates the user's subversion credentials (if
any) from the canonical emerge-subversion location to the
sandboxed environment ${HOME} directory to fix the 32-bit
clang/asan build when bootstrapping the chroot from
sources.

BUG=None
TEST=Build chroot with --bootstrap from behind
proxy

Change-Id:  I0ed3bbf6d2d3eadcd5616bc1fe39c5be966b13dd
Reviewed-on: https://gerrit.chromium.org/gerrit/34293
Commit-Ready: paul drews <paul.drews@intel.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: paul drews <paul.drews@intel.com>