chromeos-init: Fix USE=nfs support to deal with encrypted stateful
The init script now includes commit=600 on the mount, but this can't be used
with tmpfs, so update it.

Note: this is really fragile. We should instead consider adding an option
in /etc/default/stateful or similar, to select the required behaviour.
However this would add access to a file outside /sbin to the start-up path,
so I'm not sure it is desirable.

BUG=chrome-os-partner:15192
TEST=manual
Boot with non-verified firmware boot, and USE=nfs flag set for chromeos-init.
See that the stateful partition is mounted OK and the device boots to a
prompt.

Change-Id: If484e9a70a1e62284c230ca3dcaa87c95b68f2a7
Reviewed-on: https://gerrit.chromium.org/gerrit/35709
Reviewed-by: Anush Elangovan <anush@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>