chromeos-base/chromeos: depend on sys-apps/less
We explicitly depend on 'less' so that we get it in production images.
Traditionally we've had to use 'more', but that doesn't support
page-up when paging data from a pipe, and also it annoyingly exits
when reaching the bottom. Ideally we'd just be debugging on dev/test
images, but we often find ourselves debugging on production images.

This dependency adds about 400KiB to the image size and brings in no
further dependencies.

BUG=None
TEST=Built standard image; saw 'less' present and functional.

Change-Id: Ic038427c25a17f17e4594dff7bd10bf9ad2dad21
Reviewed-on: https://gerrit.chromium.org/gerrit/35747
Reviewed-by: Elly Jones <ellyjones@chromium.org>
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>