Add Touchbot ebuild to the test image
This patch adds the touchbot control library library to the test image
so that Chromebooks can be used to control the touchbot.  The touchbot
is used to mechanically manipulate the touchpad on a chromebook and test
it's functionality top-to-bottom.

This needs to be in the test image so that autotests can be written that
use the robot.  Additionally, they can be used by hand for developers
who want to manually test a machine.

The actual ebuild essentially does very little, just using the distutils
setup.py in the platforms/touchbot directory.

BUG=chromium-os:34993
TEST=Compile a test image for a snow machine.  When booted, the
libraries were present, and when executed they were able to communicate
with the robot correctly.  The same was done on a Lumpy
CQ-DEPEND=I469148496f85a626f60927a424f7b50f5f6863e0

Change-Id: I3c24d6d49e5d1c5159f70dfcd66288add55cb689
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/36226
Reviewed-by: Mike Frysinger <vapier@chromium.org>