Ensure that the NSS Internal Module is default module for operations.
NSS defaults to preferring the most recently added module when
attempting to use algorithm that it doesn't have explicit defaults for.
This was causing NSS to call into Chaps to perform SHA-256/SHA-512
operations, as well as constantly polling the TPM for session
information for a number of other algorithms (such as PKBDF2, as used by
sync, or ECC, as used by OBC)

This merges in the patches from
https://bugzilla.mozilla.org/show_bug.cgi?id=802429 , which ensure that
the internal module is the default for SHA256/SHA512, and prefers the
internal module for all other algorithms that it doesn't have explicit
defaults for.

BUG=chrome-os-partner:14707 , chromium:156237
TEST=Run crosh and enter "chaps_debug". Without the change, observe
repeated calls to C_DigestInit in /var/log/messages when accessing HTTPS
sites. With the change, observe that for non-internal HTTPS sites, the
only calls are typically C_GetSessionInfo and C_GetSlotInfo.

Change-Id: Id515024d19cefa165354e0848bb7227c5f807ae9
Reviewed-on: https://gerrit.chromium.org/gerrit/36171
Tested-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Ready: Wan-Teh Chang <wtc@google.com>
Reviewed-by: Wan-Teh Chang <wtc@google.com>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Ready: Ryan Sleevi <rsleevi@chromium.org>