xorg-server: Only display enable the main monitor on startup.
Before this change, we started X.Org in a way which is similar to
Chrome's logic. When the resolutions mismatched, this looks ugly.

As per bug chromium-os:34066, we now want the startup to only happen
on the main monitor. X.Org's initial configuration is modified
accordingly.

BUG=chromium-os:34066
TEST=by hand

Change-Id: I8b987e0b2b5c7032a11038e07f63a2fc437e7fc6
Reviewed-on: https://gerrit.chromium.org/gerrit/36300
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: Stéphane Marchesin <marcheu@chromium.org>