chromeos-bootimage: Build both RO and RW firmware images
The RO firmware has its preamble flags set to 1 in the vblocks. This
allows it to boot straight through to the kernel without jumping to RW
firmware first (the one-stop optimization).

The RW firmware is identical except that it has its preamble flags to 0.

Since cros_bundle_firmware has an easy option to specify the required
value, use this to create both firmware types.

Output images will be called:

   image-${board}.ro.bin
   image-${board}.rw.bin

Non-verified images are still build also. So for example, snow will now
have:

   image-snow.ro.bin
   image-snow.rw.bin
   nv_image-snow.bin

BUG=chromium-os:35369
BRANCH=snow
TEST=manual
$ emerge-daisy chromeos-bootimage
$ ls -l /build/daisy/firmware/*image-snow*
-rw-r--r-- 1 root root 4194304 2012-10-26 16:28 /build/daisy/firmware/image-snow.ro.bin
-rw-r--r-- 1 root root 4194304 2012-10-26 16:28 /build/daisy/firmware/image-snow.rw.bin
-rw-r--r-- 1 root root 4194304 2012-10-26 16:28 /build/daisy/firmware/nv_image-snow.bin

$ dhex .../chroot/build/daisy/firmware/image-snow.r*
Manually check that the only differences are in the fdt (1 becomes 0) and the vblock
(which changes a fair bit but that is expected).

Change-Id: I1bee06d6bb0431e406564e7e6a45c37e8f06f39d
Reviewed-on: https://gerrit.chromium.org/gerrit/36741
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>