Add bounds checking for uniform array access.
validate_uniform_parameters() now checks that the array index is
valid.  This means if an index is out of bounds, glGetUniform* will
fail with GL_INVALID_OPERATION, as it should.
It also makes checks in glUniform and glUniformMatrix redundant.
Also modified the test in glGetUniformLocation so it will return
-1 if you ask for the location of a non-existent array element.

BUG=chromium-os:145525
TEST=piglit

Change-Id: I4fb59337c8979a9759484de191712edb76c11c5d
Reviewed-on: https://gerrit.chromium.org/gerrit/36859
Tested-by: Frank Henigman <fjhenigman@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: Frank Henigman <fjhenigman@chromium.org>