vboot_reference: Install biosincludes.h for all the firmware architectures.
This file is needed by sysincludes.h which is included by vboot_api.h. Rather
than try to recreate it in the host firmware and keep the two in sync, we can
just install the one vboot uses so we can include that one.

BUG=chrome-os-partner:8339
TEST=Built on link, lumpy and snow.

Change-Id: I288395f359ad2b403d27f914472dac5000f391be
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/36887
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>