cros-workon: handle failed builds better
If a build fails, it could be for a variety of reasons.  Since one of them
could be file corruption (such as running out of disk space), throw away
the build dir.  That way the next run will start fresh and should be able
to recover automatically.

We don't throw the dir away immediately -- we rename it with a "failed"
suffix.  If a dev wants to investigate, they can use that.  On the next
run, the code will automatically delete the failed dir.

BUG=chromium-os:35958
TEST=emerging power_manager with an error in the source showed it backup the dir and then remove it on next run

Change-Id: I95e990a2f22a7a2019e99027860e2f6353b1dbba
Reviewed-on: https://gerrit.chromium.org/gerrit/37154
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>