alsa-utils: backport USE flag optimizations from upstream
These changes are taken straight from Gentoo, so nothing to worry about
when we upgrade to a newer version.

This doesn't change the files that are installed by default, but it does
allow use to disable (unused) utils that pull in other packages.  But we
will do that in a follow up commit.

BUG=chromium-os:36114
TEST=`emerge-daisy alsa-utils` installs same set of files

Change-Id: Ibd9683b8e25863e50725eca9ad804d6b6c465ebe
Reviewed-on: https://gerrit.chromium.org/gerrit/37473
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>