protofiles: move to dev/test images
We need the .proto files at build time, not runtime.  So move the package
out of RDEPEND and into DEPEND for Chrome.  This allows us to control the
images to merge the package, restricting it to dev/test images.

While we're tweaking speech_synthesis, move libresample to DEPEND too.
That package only provides a static library, so it can skip RDEPEND.

BUG=None
TEST=build_image no longer installs protofiles into base image, but does into dev/test

Change-Id: I50a9cdd230651b637bac92bb701076bf8fb73f7a
Reviewed-on: https://gerrit.chromium.org/gerrit/37772
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>