Expand CA certificate set.
The previous CA certificate set was overly restrictive and may be
incompatible with future Google production certificate changes.

This change replaces it with the standard certificate set for connecting
to Google. (Note that this set is not sufficient for connecting to hosts
on the Internet in general.)

Change-Id: I38cae5e4fd3529006eee6498d1edcd33c1ec95db
BUG=chrome-os-partner:14221
TEST=Manually tested with curl from the command line.
Reviewed-on: https://gerrit.chromium.org/gerrit/37757
Reviewed-by: Darin Petkov <petkov@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Adam Langley <agl@chromium.org>
Commit-Ready: Darin Petkov <petkov@chromium.org>
Tested-by: Darin Petkov <petkov@chromium.org>