Add a depthcharge USE flag to the chromeos-bootimage ebuild.
Enable building a depthcharge based firmware image without having to modify
chromeos-bootimage manually. This isn't the best possible solution because it
adds complexity to an already complex system, but it gets the job done without
having to wait for the firmware build process to be reworked. The additional
complexity is minimal, and is patterned after support for running with memtest.

BUG=chrome-os-partner:8339
TEST=Built and booted depthcharge and U-Boot on Link by adding and removing
the depthcharge USE flag.
BRANCH=None

Change-Id: Iad15a32b59c2fe53cdff873b6b805994d3399c9c
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/38104
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>