HACK: Add in spring USE flag detection for building spring EC
Following along the lines of 4e6bf55 we need to be able to build a spring ec
when the spring USE flag is set. This should be removed once the ec build
infrastructure is adjusted to enable building multiple ec variants for daisy
variants.

BUG=chrome-os-partner:14322
TEST=emerge-daisy_spring chromeos-ec and check for build to be spring

Change-Id: I47d561b96d58c16698756300d4a07fb9b6f6d6a6
Reviewed-on: https://gerrit.chromium.org/gerrit/38148
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Ready: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>