cros-kernel2.eclass: Update exynos5_defconfig
This is a bit of a WIP but is what I'm using for my upstream effort.
This is mostly just a 'make savedefconfig' of 'prepareconfig
chromeos-exynos5' but with a few extra bits pulled in from the current
linuxnext exynos_defconfig.

BUG=None
TEST=Booted upstream kernel on Snow to a tty (moved tty-serial to
be the startup script).

Change-Id: Ida35407f3ea7c6f8185d539c4568b3d9c6477424
Reviewed-on: https://gerrit.chromium.org/gerrit/38429
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>