x11-base/xorg-server: Rework per-Randr CRTC pointer scaling
This is a modification of
https://gerrit.chromium.org/gerrit/#/c/34344/7 . Credit to Dan Kurtz
for this.

Scale pointer motion based on the pixel density of the current "RandR" crtc
and whether or not it's internal.

This implementation is very hacky since it directly accesses XRandR methods
and datastructures from within the X server proper.  I'm not sure what
the proper way to solve this is.

BUG=chromium-os:30822
TEST=Tested on Link, Lucas w/ external display that pointer scaled
properly in each.

Change-Id: I6419b7f0a53d4411b4aa684873c0301de690a041
Reviewed-on: https://gerrit.chromium.org/gerrit/38257
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>