x11-drivers/xf86-input-evdev: Add SYN_DROPPED handling
If an evdev client cannot consume evdev events in its queue fast enough, the
evdev kernel driver will enqueue a SYN_DROPPED event and clear the queue
once the client's queue is full. The result is that the X driver will be out
of sync with respect to the kernel driver state. The patch tries to handle the
SYN_DROPPED event by retrieving the kernel driver's state. Retrieving this
state is inherently non-atomic, since it requires a sequence of ioctls. We use
a simple before and after time stamping approach to deal with the race
condition between partially syncing state and any potentially stale events that
arrive during synchronization.

BUG=chromium-os:35291
TEST=on device; run the platform_EvdevSynDropTest test
  * Execute the test script and see if it passes the test.
     cd /usr/local/autotest
     bin/autotest tests/platform_EvdevSynDropTest/control

Change-Id: Id6aa3ad9a3708563181553198b0264052e0da9dd
Reviewed-on: https://gerrit.chromium.org/gerrit/37987
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Chung-yih Wang <cywang@chromium.org>
Tested-by: Chung-yih Wang <cywang@chromium.org>