Touch: Add a location for Lumpy linearity data.
The touchpad non linearity filter uses a data file to store all the
accumulated error measurements.  There is a string xinput property that
stores where to look for this data file.  You can't change string
properties, so you have to set them at initialization.  This CL adds a
location value to Lumpy Chromebooks so that the filter can find its
data.  The filter is still disabled by default, but this way it will be
possible to enable it, or else it would not have any data to use.

BUG=chromium-os:32271
TEST=built a test image for Lumpy with this and the parameter was set
correctly, and everything else worked as expected

Change-Id: I1902d5aa62e2e214d60d99157c5845e359899aa8
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39005
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>