xf86-video-intel: Split framebuffer and flip crtcs.
Try to allocate a separate framebuffer for each CRTC and flip those
directly with the back-buffer of updated drawables. This avoids
blitting content when the drawable size matches the crtc size.

When CopyRegion is used top copy content into or out of the screen
buffer, it may be redirected to exactly one scanout buffer. This is
enough to support our one-drawable-per-crtc case, but not sufficient
in general.

This change also moves page flip state that was previously centralized
in the intel_mode structure into the intel_crtc structure associated
with the reference crtc for a flip. This allows flips to pending on
multiple pipes simultaneously.

Tested with the WebGL aquarium and space rocks demos, a Youtube
video and WebKit poster circle demo. Added and removed displays while
active and while suspended.

BUG=chromium-os:35796
TEST=Connect external monitor. Rapidly drag window horizontally,
     look for tearing.

Change-Id: Ia330e38a4a92d4e8899aac9dd841fbad13408ba4
Reviewed-on: https://gerrit.chromium.org/gerrit/39081
Commit-Ready: Chris Wolfe <cwolfe@chromium.org>
Reviewed-by: Chris Wolfe <cwolfe@chromium.org>
Tested-by: Chris Wolfe <cwolfe@chromium.org>