chromeos-base: re-enable cursor on login shells
The vt cursor was disabled to avoid messing with boot splash display
(which is perfectly fine).  Unfortunately, it also means when you log
into a vt, you no longer have a cursor.  Add a login shell fragment
to automatically turn it back on on the active vt.

BUG=chromium-os:22879
TEST=boot & log into vt2 on alex & get a cursor
TEST=run crosh & launch a shell & things work normally

Change-Id: Iab2be3f077f19221622340a67e50e859fe8172de
Reviewed-on: https://gerrit.chromium.org/gerrit/39961
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>