Add AP userspace scripts to tweak lightbar colors
We have yet another tweak for the lightbar, but we don't want to update the
EC. This CL (along with one in src/platform/ec) adds an init script that
runs on the AP at every boot and pokes the EC to modify the lightbar
settings. We have to run it at every boot because the EC will hibernate
after the AP has been off (not suspended) for an hour on battery power and
will lose its settings.

BUG=chrome-os-partner:16827
BRANCH=link
TEST=manual

Build the image for Link, install, reboot.

Run "ectool lightbar params". The output should match what's in
/usr/share/ec/lightbar_params.txt

Change-Id: I4ccdc79adef70a4735ac07ae0380d7f34dfbf071
Reviewed-on: https://gerrit.chromium.org/gerrit/39977
Commit-Queue: Bill Richardson <wfrichar@chromium.org>
Tested-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>