Remove unneeded Chrome patches from files dir.
Originally, when these patches were removed from the Chrome ebuild,
we left the old patches in place so that they could still be referenced
by the stable ebuild. As soon as the Chrome bot has revved the stable
ebuild and removed all references to these patches, we can push this patch
as well and remove the patches from the files dir.

BUG=chromium-os:34310
TEST=Launch full build with and without ninja.

Change-Id: I443dbd5084c553d083be782ca99a37b2606e4dbf
Reviewed-on: https://gerrit.chromium.org/gerrit/40228
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>