cros-firmware: fix image checksum names
It looks like the new FIRMWARE_MAIN_RW_SUM value is defined as
FIRMWARE_MAIN_RW_IMAGE_SUM, however keeping it in line with the other
nomenclature we should drop the IMAGE. Also adjusted the eclass-variables
which seemed to not match the code.

BUG=None
TEST=emerge-daisy chromeos-firmware-daisy
Verify that no checksum warning is shown for the FW RW image.

Change-Id: Iac6dc921edb0375f5fb877ec1539b5094f11f4dc
Reviewed-on: https://gerrit.chromium.org/gerrit/41158
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>