xf86-video-intel: Fix issues when switching primary display
When Chrome switches the primary display it rapidly moves and resizes
its windows without reconfiguring the crtcs. This exposes two race
conditions: a window may request a page flip with a mis-sized back
buffer, and a page flip may be requested while one is already pending
on the crtc.

This change tweaks the video driver to fall back to the blit path in
both of these race conditions. That appears to resolve the freeze,
crash and corruption problems. The I830DRI2CopyRegion code on the
blit path will merge the scanouts if necessary, but otherwise just
blits into the split framebuffer to avoid resetting the crtcs.

BUG=chromium:168950
TEST=Connect two monitors to a stumpy, log in, hold down Alt-F4 to
     rapidly switch primary displays.

Change-Id: I9cf522a430162b8f24cc8820958691eb85687163
Reviewed-on: https://gerrit.chromium.org/gerrit/41208
Commit-Queue: Chris Wolfe <cwolfe@chromium.org>
Reviewed-by: Chris Wolfe <cwolfe@chromium.org>
Tested-by: Chris Wolfe <cwolfe@chromium.org>