Introduce embedded CrOS DEPEND and RDEPEND
Organize the CrOS dependencies so we can build a very small CrOS
image for embedded environments. With the cros_embedded useflag
we reduce our dependencies and switch to busybox. This sets the
framework to add just the required packages and still use the
regular build workflow

Drop superiotool since we nolonger use it and it was x86 only

BUG=none
TEST=./build_packages --board=daisy-drm --nowithautotest --nowithfactory
./build_image --board=daisy-drm base --noenable_rootfs_verification
and boot resulting image on daisy
trybot with daisy-paladin, mario-paladin, daisy-full

Change-Id: Id50eb2e790df9a90497866555a4fee1b113d26ec
Reviewed-on: https://gerrit.chromium.org/gerrit/40153
Commit-Queue: Anush Elangovan <anush@google.com>
Reviewed-by: Anush Elangovan <anush@google.com>
Tested-by: Anush Elangovan <anush@google.com>