Fail to compile in case of bad uniform access.
An assertion in fs_visitor::remove_dead_constants() would fail on code like
this, which accesses a non-existent uniform:
  uniform vec u[1];
  ...
  a += u[0] + u[1];

The code is checked for that error, but apparently only before transformations
like loop unrolling.  So the following code would get transformed to the same
as above and then the assertion would fail:
  for (int i = 0; i < 2; ++i) a += u[i];

This patch changes the assertion failure to a compile failure.
It also adds initialization for a variable that previously had none but is
tested by another assertion in the same function.

Upstream bug: https://bugs.freedesktop.org/show_bug.cgi?id=59429

BUG=chromium:169054
TEST=run test.c from bug report and see shader link error (cannot test with chrome now because it works around the bug)

Change-Id: Icbe560f704abeed87e36fa00c89062c3197c9add
Reviewed-on: https://gerrit.chromium.org/gerrit/41548
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Frank Henigman <fjhenigman@chromium.org>
Commit-Queue: Frank Henigman <fjhenigman@chromium.org>