imagemagick: set prefix to /usr/local
We copied imagemagick to our tree for some cross-compiling fixes and
to set the runtime paths to /usr/local.  Since the fixes are upstream,
that leaves the /usr/local hack, so move that to bashrc so we do not
have to modify the ebuild itself.

BUG=chromium-os:31039
TEST=`emerge-x86-alex imagemagick` and running `convert` in the test image works
TEST=`cbuildbot {arm,amd64,x86}-generic-full` work
TEST=build_image for factory and verified the .la files were in the image

Change-Id: I62bf6369f13cf18b4baee4aaa158a636d50b4ac0
Reviewed-on: https://gerrit.chromium.org/gerrit/41703
Reviewed-by: David James <davidjames@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>