nspr/nss: mask new versions
The newer nspr/nss are breaking vmtests.  So we want to mask them out.

The older packages have a bug where they build against the local libs
but end up using the system libs to run some helpers.  This triggers a
crash.  So fix that in the process so the downgrade works.

BUG=chromium-os:38334
TEST=`./update_chroot` works to downgrade nspr/nss

Change-Id: Ibf48a752427c595699aab7cebbdca044fe6eaf92
Reviewed-on: https://gerrit.chromium.org/gerrit/42233
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris Wolfe <cwolfe@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>