chromeos-chrome: enable std::vector protection
This adds -D__google_stl_debug_vector=1 to the Chrome build to enable
bounds-checking when using std::vector []-operator. Performance tests
(pagecycler) on Chrome built with this show no measurable changes. It
gains us protections against out-of-array-bounds accesses, which has
been a real exploit path in the past on Chrome.

Once this has been checked more widely on Canaries, we can add this flag
to the global builds via the sysroot wrapper. This CL can also serve as
a single CL to backport this feature to Chrome in R25.

BUG=chromium-os:38348
TEST=Chrome with pagecycler runs without performance changes

Change-Id: Iefd4cb7eb1be872cde77b9a312a748523a6fbe2d
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/42402
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>