Adding support for profiling to cros-workon.eclass
Added in support for a 'profiling' USE flag that is translated to a MODE
variable value and added cros-workon_src_install to install the generated
coverage report into the appropriate directory,
/usr/share/profiling/<package_name>/lcov.

This also adds using the function to the power_manager ebuild since I needed it
for testing. This function is written so that if common.mk has not been upgraded
it works fine, so 42330 and this CL are not dependent on each other.

CQ-DEPENDS=I882fa7a7496db52a72bb8c5409b3715490d95879
BUG=chromium-os:37854
TEST=<TBD>
     Copied common.mk from gerrit.chromium.org/gerrit/42330/ to power_manager
     for testing.
     `USE=profiling emerge-$BOARD power_manager`
     `FEATURES=test USE=profiling emerge-$BOARD power_manager`
     Tested that the second version installed lcov output in
     /build/../usr/share/profiling/power_manager/lcov/
     Tested that coverage files were not installed to an image

Change-Id: I5c678d4325785d996a599fa099b0efb1c6b1ef49
Reviewed-on: https://gerrit.chromium.org/gerrit/42339
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>