mesa: clean up cross-compile logic a bit [try 2]
Our cross-compile changes, while an improvement, still don't solve
the entire problem.  The current logic only works because:
 - the generated code tends to be simple
 - our sdk currently runs on x86-compatible CPU's
 - we install 32bit multilibs in the sdk

If either of these assumptions are voided, the build of mesa fails.
Rework the code to behave like the kernel's asm-offsets -- we parse
custom assembly output and rewrite it into a header.  Now we do not
have to execute anything.

This is an update from CL:15109.  That one was reverted because it
broke amd64 targets.  I've double verified that doesn't happen with
this version.  I also had to rebase onto the latest mesa tree since
things have changed a lot.

BUG=chromium-os:36866
TEST=`emerge-x86-alex mesa` still works
TEST=`emerge-stumpy mesa` still works
TEST=`emerge-x32-generic mesa` now works
TEST=hand inspection of generate matypes.h results in same defines/values for x86 & amd64 systems

Change-Id: I036ae559f0b4f4ad38b3d62259ebe797a4369dba
Reviewed-on: https://gerrit.chromium.org/gerrit/42500
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>