Update all of the common.mk based package ebuilds to support profiling
This CL adds calling the src_install function from the cros_workon.eclass in the
ebuilds of all of the packages that we have the use common.mk. This enables
profiling support once the associated common.mk has been updated to the newest
version. This change and the common.mk update are designed to be able to be done
independent of each other.

power_manager has already been updated, since it was the test package for this
process and common-mk is not built, so doesn't need this.

x11-drivers/touchpad-tests is not being updated at this time, since it is
currently broken and I can't test my ebuild change. The bug for tracking fixing
it is crosbug.com/38449 and I will update common.mk once it is fixed.

chromeos-base/chromeos-debugd is not being updated at this time, because it puts
its common.mk in a non-top level location, which breaks a bunch of the eclass
stuff. I will contact the owner of this pacakge about this.

Packages updated:
  - chromeos-base/chaps
  - chromeos-base/chromeos-cryptohome
  - chromeos-base/chromeos-installer
  - chromeos-base/crash-reporter
  - chromeos-base/cros-disks
  - chromeos-base/cros_boot_mode
  - chromeos-base/libevdev
  - chromeos-base/mtpd
  - chromeos-base/permission_broker
  - chromeos-base/verity
  - chromeos-base/vpn-manager
  - chromeos-base/wimax_manager
  - dev-util/xnu_quick_test
  - sys-apps/bootcache
  - sys-apps/rollog

BUG=chromium-os:37854
TEST=For each package:
     `MODE=profiling cros_workon_make --board=link <package>`
     `MODE=profiling cros_workon_make --board=link <package> --test`
     `cros_workon_make --board=link <package>`
     `cros_workon_make --board=link <package> --test`
     Repeated these with emerge-link <package> and FEATURES=test as need

Change-Id: I206e3c337b9a5cbde22933341899bb0122e19d11
Reviewed-on: https://gerrit.chromium.org/gerrit/42743
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>